3_Agent_to_Deploy_Assets/cdk-python/cdk.out/asset.b0e3113168a462fd399dce3b31b045d09f97e57b82c2b2dc97f472b2811ffbe5/requirements.txt
PyJWT==2.8.0
requests==2.31.0
cryptography==41.0.7